+++ packages/core/src/http-client/http-client.ts
@@ -13,7 +13,7 @@ import { getAgentConfig } from './http-agent';
 import {
   DestinationHttpRequestConfig,
   ExecuteHttpRequestFn,
-  HttpRequest,
+  HttpRequest, HttpRequestAndResponse,
   HttpRequestConfig,
   HttpResponse
 } from './http-client-types';
@@ -81,11 +81,11 @@ export function addDestinationToRequestConfig<T extends HttpRequestConfig>(
  * @param executeFn - A function that can execute an [[HttpRequestConfig]].
  * @returns A function expecting destination and a request.
  */
-export function execute(executeFn: ExecuteHttpRequestFn) {
+export function execute<ReturnT>(executeFn: ExecuteHttpRequestFn<ReturnT>) {
   return async function <T extends HttpRequestConfig>(
     destination: Destination | DestinationNameAndJwt,
     requestConfig: T
-  ): Promise<HttpResponse> {
+  ): Promise<ReturnT> {
     const destinationRequestConfig = await buildHttpRequest(
       destination,
       requestConfig.headers
@@ -125,6 +125,16 @@ export async function buildAxiosRequestConfig<T extends HttpRequestConfig>(
  * @returns An [[HttpResponse]].
  */
 export const executeHttpRequest = execute(executeWithAxios);
+
+/**
+ * Builds a [[DestinationHttpRequestConfig]] for the given destination, merges it into the given requestConfig
+ * and executes it (using Axios). The request is also included in the object returned.
+ *
+ * @param destination - A destination or a destination name and a JWT.
+ * @param requestConfig - Any object representing an HTTP request.
+ * @returns An [[HttpRequestAndResponse]].
+ */
+export const executeHttpRequestReturnRequestAndResponse = execute(executeWithAxiosReturnRequestAndResponse);
 
 function buildDestinationHttpRequestConfig(
   destination: Destination,
@@ -183,8 +193,19 @@ function merge<T extends HttpRequestConfig>(
   };
 }
 
+function mergeRequestWithAxiosDefaults(request: HttpRequest): HttpRequest{
+  return { ...getAxiosConfigWithDefaults(), ...request };
+}
+
 function executeWithAxios(request: HttpRequest): Promise<HttpResponse> {
-  return axios.request({ ...getAxiosConfigWithDefaults(), ...request });
+  return axios.request(mergeRequestWithAxiosDefaults(request));
+}
+
+function executeWithAxiosReturnRequestAndResponse(request: HttpRequest): Promise<HttpRequestAndResponse> {
+  const merged = mergeRequestWithAxiosDefaults(request);
+  return axios.request(merged).then(
+    res => ({ request: merged, response: res })
+  );
 }
 
 /**