+++ packages/core/src/odata-common/request-builder/action-function-import-request-builder-base.ts
@@ -6,6 +6,7 @@ import {
 import { MethodRequestBuilder } from '../request-builder/request-builder-base';
 import { ODataFunctionImportRequestConfig } from '../request/odata-function-import-request-config';
 import { ODataActionImportRequestConfig } from '../../odata-v4';
+import { HttpRequestAndResponse } from '../../http-client';
 
 /**
  * Create OData request to execute a action or function import.
@@ -43,9 +44,23 @@ export abstract class ActionFunctionImportRequestBuilder<
     destination: Destination | DestinationNameAndJwt,
     options?: DestinationOptions
   ): Promise<ReturnT> {
+    return this.executeRaw(destination, options)
+      .then(({ response }) => this.responseTransformer(response.data));
+  }
+
+  /**
+   * Execute request and return the request and the raw response.
+   *
+   * @param destination - Destination to execute the request against
+   * @param options - Options to employ when fetching destinations
+   * @returns A promise resolving to an [[HttpRequestAndResponse]].
+   */
+  async executeRaw(
+    destination: Destination | DestinationNameAndJwt,
+    options?: DestinationOptions
+  ): Promise<HttpRequestAndResponse>{
     return this.build(destination, options)
-      .then(request => request.execute())
-      .then(({ data }) => this.responseTransformer(data));
+      .then(request => request.executeRaw());
   }
 }