+++ packages/core/src/odata-common/request-builder/get-request-builder-base.ts
@@ -4,6 +4,8 @@ import { Selectable } from '../selectable/selectable';
 import { ODataGetAllRequestConfig } from '../request/odata-get-all-request-config';
 import { MethodRequestBuilder } from '../request-builder/request-builder-base';
 import { ODataGetByKeyRequestConfig } from '../request';
+import { Destination, DestinationNameAndJwt, DestinationOptions } from '../../connectivity/scp-cf/destination';
+import { HttpRequestAndResponse } from '../../http-client';
 
 /**
  * Base class for the get all and get by key request builders.
@@ -46,6 +48,21 @@ export abstract class GetRequestBuilder<
   ): this {
     this.requestConfig.selects = variadicArgumentToArray(first, rest);
     return this;
+  }
+
+  /**
+   * Execute request and return the request and the raw response.
+   *
+   * @param destination - Destination to execute the request against
+   * @param options - Options to employ when fetching destinations
+   * @returns A promise resolving to an [[HttpRequestAndResponse]].
+   */
+  async executeRaw(
+    destination: Destination | DestinationNameAndJwt,
+    options?: DestinationOptions
+  ): Promise<HttpRequestAndResponse>{
+    return this.build(destination, options)
+      .then(request => request.executeRaw());
   }
 }