+++ test-packages/type-tests/test/http-client.spec.ts
@@ -1,4 +1,10 @@
-import { executeHttpRequest } from '@sap-cloud-sdk/core';
+import {
+  executeHttpRequest,
+  executeHttpRequestReturnRequestAndResponse
+} from '@sap-cloud-sdk/core';
 
 // $ExpectType Promise<HttpResponse>
 executeHttpRequest({ url: 'https://example.com' }, { method: 'get' });
+
+// $ExpectType Promise<HttpRequestAndResponse>
+executeHttpRequestReturnRequestAndResponse({ url: 'https://example.com' }, { method: 'get' });